Extracted overlay to file
