+++ lib/assets/javascripts/cartodb3/data/feature-definition-model.js
@@ -81,6 +81,7 @@ var FeatureDefinitionModel = Backbone.Model.extend({
     var attrs = _.pick(this.toJSON(), _.difference(columns, BLACKLISTED_COLUMNS));
 
     this._getQueryRowModel().save(attrs);
+    this.trigger('save');
   },
 
   _getQueryRowModel: function () {
