+++ lib/assets/javascripts/cartodb3/deep-insights-integrations.js
@@ -1,6 +1,4 @@
 var _ = require('underscore');
-var $ = require('jquery');
-var Backbone = require('backbone');
 var linkLayerInfowindow = require('./deep-insights-integration/link-layer-infowindow');
 var linkLayerTooltip = require('./deep-insights-integration/link-layer-tooltip');
 var LegendManager = require('./deep-insights-integration/legend-manager');
@@ -9,63 +7,6 @@ var WidgetsNotifications = require('./widgets-notifications');
 var AnalysisOnboardingLauncher = require('./components/onboardings/analysis/analysis-launcher');
 var VisNotifications = require('./vis-notifications');
 var FeatureDefinitionModel = require('./data/feature-definition-model');
-
-// TODO: Destroy
-var Overlay = Backbone.View.extend({
-  className: 'CDB-Overlay',
-  type: 'custom',
-  template: '<p class="js-edit-feature">Edit</p>',
-
-  events: {
-    'click .js-edit-feature': '_onEditButtonClicked'
-  },
-
-  initialize: function (options) {
-    _.bindAll(this, '_onEditButtonClicked');
-
-    if (!options.map) throw new Error('map is required');
-    this._map = options.map;
-
-    // TODO: Make this customizable hideOnPan, hideOnZoom
-    this._map.on('change:center', this.hide, this);
-    this._map.on('change:zoom', this.hide, this);
-
-    this._position = {
-      x: 0,
-      y: 0
-    };
-  },
-
-  render: function () {
-    this.$el.html(this.template);
-    return this;
-  },
-
-  hide: function () {
-    this.$el.hide();
-    $(document.body).off('click', this.hide.bind(this));
-  },
-
-  show: function () {
-    this.$el.css({
-      width: '20px',
-      top: this._position.y,
-      left: this._position.x
-    });
-    this.$el.show();
-
-    $(document.body).on('click', this.hide.bind(this));
-  },
-
-  setPosition: function (position) {
-    this._position = position;
-  },
-
-  _onEditButtonClicked: function (event) {
-    event.stopPropagation();
-    this.trigger('editFeature');
-  }
-});
 
 /**
  * Integration between various data collections/models with cartodb.js and deep-insights.
@@ -86,6 +27,7 @@ var F = function (opts) {
   if (!opts.mapModeModel) throw new Error('mapModeModel is required');
   if (!opts.configModel) throw new Error('configModel is required');
   if (!opts.editorModel) throw new Error('editorModel is required');
+  if (!opts.editFeatureOverlay) throw new Error('editFeatureOverlay is required');
 
   this._diDashboard = opts.deepInsightsDashboard;
   this._analysisDefinitionNodesCollection = opts.analysisDefinitionNodesCollection;
@@ -101,6 +43,7 @@ var F = function (opts) {
   this._configModel = opts.configModel;
   this._widgetDefinitionsCollection = opts.widgetDefinitionsCollection;
   this._editorModel = opts.editorModel;
+  this._editFeatureOverlay = opts.editFeatureOverlay;
 
   this._layerDefinitionsCollection.each(this._linkLayerErrors, this);
 
@@ -169,22 +112,10 @@ var F = function (opts) {
 
   opts.mapModeModel.on('change:mode', this._onMapModeChanged, this);
 
-  var vis = this._vis();
-
-  this._editFeatureOverlay = new Overlay({
-    map: vis.map
-  });
-
-  this._editFeatureOverlay.hide();
-
-  vis.addCustomOverlay(this._editFeatureOverlay);
-
   this.visMap().on('featureClick', this._onFeatureClicked, this);
 };
 
 F.prototype._onFeatureClicked = function (event) {
-  this._editFeatureOverlay.off('editFeature');
-
   var layerId = event.layer.id;
   var featureId = event.feature.cartodb_id;
   var position = event.position;
@@ -193,10 +124,8 @@ F.prototype._onFeatureClicked = function (event) {
   if (layerDefinitionModel && layerDefinitionModel.getAnalysisDefinitionNodeModel().get('type') === 'source') {
     this._editFeatureOverlay.setPosition(position);
     this._editFeatureOverlay.show();
-
-    this._editFeatureOverlay.on('editFeature', function () {
-      this._editFeatureOverlay.hide();
-
+    this._editFeatureOverlay.off('editFeature');
+    this._editFeatureOverlay.once('editFeature', function () {
       var featureDefinition = new FeatureDefinitionModel({
         id: featureId
       }, {
@@ -206,13 +135,11 @@ F.prototype._onFeatureClicked = function (event) {
 
       featureDefinition.fetch({
         success: function () {
+          // Exit editing feature mode to re-enter this mode again
           if (this._mapModeModel.isEditingFeatureMode()) {
-            var xfeatureDefinition = this._mapModeModel.getFeatureDefinition();
-            xfeatureDefinition.set(featureDefinition.toJSON());
-            this._onMapModeChanged(this._mapModeModel);
-          } else {
-            this._mapModeModel.enterEditingFeatureMode(featureDefinition);
+            this._mapModeModel.enterViewingMode();
           }
+          this._mapModeModel.enterEditingFeatureMode(featureDefinition);
         }.bind(this),
         error: function () {
           // TODO: Handle errors here!
@@ -247,8 +174,6 @@ F.prototype._onMapModeChanged = function (mapModeModel) {
     if (!geometry) {
       throw new Error("couldn't get geometry for feature of type " + featureDefinition.get('type'));
     }
-
-    this._bindGeometryToFeatureDefinition(geometry, featureDefinition);
   }
 
   // EDITING FEATURES
@@ -257,7 +182,13 @@ F.prototype._onMapModeChanged = function (mapModeModel) {
     var geojson = JSON.parse(featureDefinition.get('the_geom'));
 
     geometry = map.editGeoJSONGeometry(geojson);
+  }
+
+  if (featureDefinition && geometry) {
     this._bindGeometryToFeatureDefinition(geometry, featureDefinition);
+    featureDefinition.on('save', function () {
+      this._invalidateMap();
+    }, this);
   }
 };
