+++ lib/assets/javascripts/cartodb3/editor.js
@@ -37,6 +37,7 @@ var ImporterManager = require('./components/background-importer/background-impor
 var BackgroundPollingModel = require('./data/editor-background-polling-model');
 var StyleManager = require('./editor/style/style-manager');
 var DeepInsightsIntegrations = require('./deep-insights-integrations');
+var EditFeatureOverlay = require('./deep-insights-integration/edit-feature-overlay');
 var Notifier = require('./components/notifier/notifier');
 var MetricsTracker = require('./components/metrics/metrics-tracker');
 var FeedbackButtonView = require('./editor/feedback/feedback-button-view');
@@ -249,6 +250,12 @@ var mapCreation = function () {
 
     var styleManager = new StyleManager(layerDefinitionsCollection, vis.map);
 
+    var editFeatureOverlay = new EditFeatureOverlay({
+      map: vis.map
+    });
+    editFeatureOverlay.hide();
+    vis.addCustomOverlay(editFeatureOverlay);
+
     var deepInsightsIntegrations = new DeepInsightsIntegrations({
       onboardings: onboardings,
       userModel: userModel,
@@ -265,7 +272,8 @@ var mapCreation = function () {
       stateDefinitionModel: stateDefModel,
       mapModeModel: mapModeModel,
       configModel: configModel,
-      editorModel: editorModel
+      editorModel: editorModel,
+      editFeatureOverlay: editFeatureOverlay
     });
 
     // Expose things after Map initialization
